move pickers + rename base + fix deselection filtering
